+++ src/main/java/com/example/dditlms/domain/repository/MemberRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByMemberId(String memberId);
-    Optional<Member> findByUserNumber(int userNumber);
+    Optional<Member> findByUserNumber(long userNumber);
 }
